+++ middleware_create.go
@@ -1,8 +1,7 @@
-package command
+package gwf
 
 import (
 	"fmt"
-	"github.com/RobyFerro/go-web-framework"
 	"io/ioutil"
 	"path"
 	"path/filepath"
@@ -10,32 +9,36 @@ import (
 	"strings"
 )
 
+// MiddlewareCreate will create a new http middleware
 type MiddlewareCreate struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
-// Command registration
+// Register this command
 func (c *MiddlewareCreate) Register() {
 	c.Signature = "middleware:create <name>" // Change command signature
 	c.Description = "Create new middleware"  // Change command description
 }
 
-// Command business logic
-func (c *MiddlewareCreate) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+// Run this command
+func (c *MiddlewareCreate) Run() {
 	var _, filename, _, _ = runtime.Caller(0)
-	splitName := strings.Split(strings.ToLower(args), "_")
+
+	splitName := strings.Split(strings.ToLower(c.Args), "_")
 	for i, name := range splitName {
 		splitName[i] = strings.Title(name)
 	}
 
 	cName := strings.Join(splitName, "")
-	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "../../raw/middleware.raw"))
+	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "raw/middleware.raw"))
 
 	cContent := strings.ReplaceAll(string(input), "@@TMP@@", cName)
-	cFile := fmt.Sprintf("%s/%s.go", gwf.GetDynamicPath("app/http/middleware"), strings.ToLower(args))
+	cFile := fmt.Sprintf("%s/%s.go", GetDynamicPath("app/http/middleware"), strings.ToLower(c.Args))
+
 	if err := ioutil.WriteFile(cFile, []byte(cContent), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("\nSUCCESS: Your %s middleware has been created at %s\n", cName, cFile)