+++ show_route.go
@@ -1,28 +1,32 @@
-package command
+package gwf
 
 import (
-	"github.com/RobyFerro/go-web-framework"
-	"github.com/olekukonko/tablewriter"
 	"os"
 	"strings"
+
+	"github.com/olekukonko/tablewriter"
 )
 
+// ShowRoute will shows all routes registerer in your Go-Web application
 type ShowRoute struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
+// Register this command
 func (c *ShowRoute) Register() {
 	c.Signature = "show:route"
 	c.Description = "Show active Go-Web routes"
 }
 
-// Show the current go-web routes
-func (c *ShowRoute) Run(sc *gwf.HttpKernel, args string, console map[string]interface{}) {
+// Run this command
+func (c *ShowRoute) Run() {
 	var data [][]string
-	routes, err := gwf.ConfigurationWeb()
+
+	routes, err := ConfigurationWeb()
 	if err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	// Parse single route
@@ -42,7 +46,7 @@ func (c *ShowRoute) Run(sc *gwf.HttpKernel, args string, console map[string]inte
 }
 
 // Show single routes
-func showSingleRoute(routes map[string]gwf.Route, data *[][]string) {
+func showSingleRoute(routes map[string]Route, data *[][]string) {
 	for _, r := range routes {
 		*data = append(*data, []string{
 			r.Method,
@@ -56,7 +60,7 @@ func showSingleRoute(routes map[string]gwf.Route, data *[][]string) {
 }
 
 // Show groups
-func showGroupRoutes(routes map[string]gwf.Group, data *[][]string) {
+func showGroupRoutes(routes map[string]Group, data *[][]string) {
 	for _, g := range routes {
 		var middleware []string
 		middleware = append(middleware, g.Middleware...)