+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "Go-Web Framework" will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.3.0-beta] - 2020-02-24
+### Changed
+- Basic type structure: controllers, services, models, and middleware.
+
+### Removed
+- Removed external configuration.
+
+### Fixed
+- Fix authentication flow in JWTAuth (other fix)
+
 ## [v0.2.2-beta] - 2020-02-19
 ### Fixed
 - Fix authentication flow in JWTAuth (other fix)